requirements.txt
certifi==2017.11.5
colorama==0.3.9
emoji==0.4.5
future==0.16.0
libkeepass==0.2.0
lxml==4.1.1
peewee==2.10.2
pycrypto==2.6.1
python-telegram-bot==8.1.1
